Fix bug when passing anything to old_canvas (= oldSwapchain)
Because the ownership of old_canvas was being passed to Canvas::new,
it was dropped near the beginning of the swapchain creation function,
probably after the .map(), which clearly was too early. Passing it as
a borrow ensures that it lives at least until the creation function
has finished, i.e. when the old swapchain can actually be destroyed.

Flimsy things, these Vulkan objects without lifetimes. Who knew! ;)

Funnily enough, I thought the issue was X11 specific, but turns out
that Wayland straight up segfaulted during the first frame with the
old buggy behaviour. One thing is for sure: sway isn't subtle when it
comes to misbehaving clients!