Refactor buffer uploads and frame locals
The previous version ended up waiting for vsync when waiting for the
temporary resource fences. In addition, the fences were a temporary
solution anyway, so I replaced them with semaphores in this one, moved
temporary buffers and buffer uploads into frame local storage (so 3
copies of everything), now that the buffer uploads could be running in
the background even after calling render. Now it seems to work
properly, since vsync wait seems to be happening when waiting on
finished_queue_fence (as intended).

Sidenote: Tracy is very nice! Wouldn't have noticed the vsync
wait falling on the wrong fences without it.