Use image indices as frame indices
Where "image index" refers to what vkAcquireNextImageKHR returns, and
"frame index" refers to the index used when accessing frame-local
resources in this crate.

I think this is more correct than the previous implementation: now,
we only reuse temporary resources if we've acquired the same image
that was rendered to, using the temporary resources from that
image. Where as before, if the image indices were out of order, we
might've used temporary resources that were still in use, since the
frame indices were just incremented every frame. Hard to test though.