Remove Gpu, add Renderer
Technically it was just a rename from Gpu to Renderer, but now the
purpose is very different: Gpu was the creator and owner of the Vulkan
Device, where now the Renderer merely holds the per-frame objects, and
only acts as a context for the actual rendering process. So where
previously the "creation/drop" hierarchy was:

Driver
  Gpu
    Canvas

Now it's:

Driver
  Device
    Canvas
      Renderer

This now allows for the per-frame objects to be created *after*
Canvas, which also means that there's actually the right amount of
them, depending on the amount of images in the swapchain.