Move mesh buffer copy over to VulkanArena for optimization
If the GPU supports device local + host visible(+coherent) memory,
Buffers are faster to write directly than via a staging buffer.

On my machine (b550/ryzen 5900x/radeon 6800 xt), with the sandbox
testing binary, the time taken by Gltf::from_gltf + Uploader::wait
went from ~100ms to ~70ms.

The copying code was moved so that only one place has to care whether
or not the asset arena (device local) is host visible, and therefore,
if the upload needs a staging buffer. It's still possible to get the
old behaviour of create_buffer by just passing in a None uploader,
then it just errors out if the buffer can't be written to without a
staging buffer.