Add vulkan memory requirement measuring
Now all the arenas can be created with exactly the required size, if
all the loaded resources are known ahead-of-time (so that they can be
measured before vulkan arena creation). These numbers can also be
exposed to the user in an options menu, for them to be able to tweak
the settings to fit their VRAM just right.