Decrease MAX_TEXTURE_COUNT to 14 to support GTX 1060
The issue is in the uniform buffer limit, which is 15.
This could also be fixed by providing the GltfFactors to the
fragment shader by some other method, but this would be much
harder. In that case though, we'd be limited by Intel again,
at ~200 sampled images if I remember correctly.