Start using VK_KHR_synchronization2, fix barriers
Spent way too long trying to figure out what was wrong with my
transfer barriers and subpass dependencies. Turns out, no issues found
with the newest version of the validation layers, it was a bug in
them. Nyegh.

At least now I'm using the new sync api and cleaned up some of the
barriers. And I certainly learned a lot, so it's not all bad.