Restructure modules, add a little bit of hierarchy
The hierarchy is based on uses: the general rule while refactoring
was, if some module is mostly referred to by just one module, move the
former module under the latter. Basically just trying to split up
parts of the codebase that do not refer to each other into different
folders. Many cases were not as clear cut, but that was the main
idea. I think this is better than the flat hierarchy there was before.