Fix projection and add coordinate system handling
Previously the rendering world space's coordinate system wasn't
well-defined, and the projection was just something that looked
correct. The new one's been crafted from the ground up to be correct
when given coordinates in Vulkan's clip-space, and I'm relatively sure
it's the right one now, given that I found a blog post that also
described it as correct in the same way*. On the other end, Scene now
has a world_space parameter which can be set to whichever coordinate
system is appropriate for the application.

* https://johannesugb.github.io/gpu-programming/setting-up-a-proper-vulkan-projection-matrix/