Refactor descriptors for indirect rendering [wip]
Next up is the actual implementation of:

- where and how are the VkDrawIndirectCommands collected
- where and how are the vkCmdDrawIndirects called

It isn't quite straightforward, because currently we're changing the
buffers for each draw, which isn't really possible when drawing
indirectly (you only get to control the vertex offset, but not the
buffer its referring to). Two initial ideas:

- make a "buffer arena" where e.g. position vertices can be appended
  onto, so that when rendering, we could just bind the "general
  position buffer" and all the offsets would be right from the
  start. It would require even more "estimating how big this buffer
  needs to be" than is currently needed with the arenas, so we'll se
  if that can be made ergonomic.
- make temporary buffers at draw time, copy over the meshes that were
  actually used, and render from that. Lots more VRAM usage, but this
  would not have the issues of the previous point, and would require
  less changes overall.

We'll see what I end up doing in the end.